AmRtpStream: change rtp stats counters to fit database integer limits

+++ core/AmRtpStream.h
@@ -472,9 +472,9 @@ public:
   void getPayloadsHistory(PayloadsHistory &ph);
 
   struct ErrorsStats {
-	  unsigned long decode_errors;
-	  unsigned long rtp_parse_errors;
-	  unsigned long out_of_buffer_errors;
+	  int decode_errors;
+	  int rtp_parse_errors;
+	  int out_of_buffer_errors;
 	  ErrorsStats(): decode_errors(0), rtp_parse_errors(0), out_of_buffer_errors(0) {}
   };
   void getErrorsStats(ErrorsStats &es);
